refactor :: dataStructureModule로 강등
